+++ test/node/features/fixtures/koa/scenarios/app.js
@@ -64,6 +64,23 @@ app.use(async (ctx, next) => {
   }
 })
 
+app.on('error', (err, ctx) => {
+  // in the "error-before-handler" test ctx.bugsnag won't exist
+  const bugsnag = ctx.bugsnag || Bugsnag
+
+  bugsnag.addMetadata('error_handler', 'before', true)
+})
+
 app.on('error', middleware.errorHandler)
+
+app.on('error', (err, ctx) => {
+  // in the "error-before-handler" test ctx.bugsnag won't exist
+  const bugsnag = ctx.bugsnag || Bugsnag
+
+  // this should not be added to events as error handlers run in the order they
+  // were added, so at this point the event has already been sent by the Bugsnag
+  // errorHandler
+  bugsnag.addMetadata('error_handler', 'after', true)
+})
 
 app.listen(80)